primitive magnetic cell
