+++ moyo/src/base.rs
@@ -1,6 +1,8 @@
+mod action;
 mod cell;
 mod error;
 mod lattice;
+mod magnetic_cell;
 mod operation;
 mod permutation;
 mod tolerance;
@@ -9,6 +11,7 @@ mod transformation;
 pub use cell::{AtomicSpecie, Cell, Position};
 pub use error::MoyoError;
 pub use lattice::Lattice;
+pub use magnetic_cell::{Collinear, MagneticCell, MagneticMoment, NonCollinear};
 pub use operation::{project_rotations, Operation, Operations, Rotation, Rotations, Translation};
 pub use permutation::Permutation;
 pub use tolerance::AngleTolerance;
