+++ moyo/src/search.rs
@@ -1,9 +1,11 @@
 mod primitive_cell;
+mod primitive_symmetry_search;
 mod solve;
 mod symmetry_search;
 
 pub use primitive_cell::PrimitiveCell;
+pub use primitive_symmetry_search::PrimitiveSymmetrySearch;
 pub use solve::{
     solve_correspondence, solve_correspondence_naive, PeriodicKdTree, PeriodicNeighbor,
 };
-pub use symmetry_search::PrimitiveSymmetrySearch;
+pub use symmetry_search::iterative_symmetry_search;
